pyproject.toml
[project]
name = "gcp-mcp-server"
version = "0.1.0"
description = "A Model Context Protocol server that provides access to Google Cloud Platform services, enabling LLMs to manage and interact with GCP resources."
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "mcp[cli]>=1.0.0",
    "google-cloud-bigquery>=3.27.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-run>=0.9.0",
    "google-cloud-artifact-registry>=1.10.0",
    "google-cloud-logging>=3.5.0",
    "python-dotenv>=1.0.0",
    "google-cloud-monitoring>=2.0.0",
    "google-cloud-compute>=1.0.0",
    "google-cloud-build>=3.0.0",
]
[[project.authors]]
name = "Enes Bol"
email = "enes2277@gmail.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gcp_mcp"]

[project.scripts]
gcp-mcp-server = "gcp_mcp.main:main"
